fix(tests): split test_unified_noop_gate.py to fix LOC ceiling violation
Split the 1214-line test_unified_noop_gate.py into 3 files to comply
with the 400-line CI block threshold:

- test_noop_gate_basic.py (539 LOC): core gate behavior + ref checking
- test_noop_gate_transition_count.py (409 LOC): transition count logic
- test_noop_gate_integration.py (262 LOC): real-DB integration tests

Shared helpers extracted to conftest.py. Updated loc_limits.yaml with
exceptions for the two files still above 400 LOC.

Consolidation: 4 original noop_gate files → 3 (partial, LOC-safe).
